+++ CHANGELOG.md
@@ -5,11 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.26](https://github.com/alloy-rs/chains/releases/tag/v0.1.26) - 2024-08-08
+
+### Other
+
+- Add acala karura mandala network ([#76](https://github.com/alloy-rs/chains/issues/76))
+
 ## [0.1.25](https://github.com/alloy-rs/chains/releases/tag/v0.1.25) - 2024-08-02
 
 ### Features
 
 - Add avalanche to the list of chains to support shanghai ([#75](https://github.com/alloy-rs/chains/issues/75))
+
+### Miscellaneous Tasks
+
+- Release 0.1.25
 
 ## [0.1.24](https://github.com/alloy-rs/chains/releases/tag/v0.1.24) - 2024-07-30